4/20 - Code Simplification, Depth Drawing, Overlay Near Finished

+++ Code/Qbert/main.cpp
@@ -2,7 +2,7 @@
 	
 int main()
 {
-	srand( time( NULL ) );
+	srand( (size_t)time( NULL ) );
 
 	StateManager game;
